+++ src/stopwatch.rs
@@ -597,16 +597,19 @@ impl Stopwatch {
     /// # Ok(())
     /// # }
     /// ```
+    #[inline]
     pub fn set_in_place(&mut self, new: Duration) {
+        self.set_in_place_at(new, Instant::now());
+    }
+
+    // TODO: add example that shows use case
+    /// Sets the total elapsed time to `new` as if the current time were
+    /// `anchor`, and without affecting whether the stopwatch is running.
+    pub fn set_in_place_at(&mut self, new: Duration, anchor: Instant) {
         let was_running = self.is_running();
         self.set(new);
         if was_running {
-            // TODO: implicit Instant::now(), meaning theres an alternate
-            // universe where i've added
-            // `fn set_in_place_at(&mut self, new: Duration, anchor: Instant)`
-            // and a corresponding `reset_in_place_at`. this could exist??
-            // but idk the use case.
-            let result = self.start();
+            let result = self.start_at(anchor);
             debug_assert!(result.is_ok());
         }
     }